requirements.txt
Django==1.5
Pyro4==4.17
South==0.7.6
amqp==1.0.9
amqplib==1.0.2
anyjson==0.3.3
beautifulsoup4==4.1.3
billiard==2.7.3.22
boto==2.2.1
celery==3.0.16
distribute==0.6.28
dj-database-url==0.2.1
django-cachebuster==0.2.1
django-celery==3.0.10
django-pylibmc-sasl==0.2.4
django-storages==1.1.4
djangorestframework==2.2.5
gensim==0.8.6
gunicorn==0.14.6
httplib2==0.7.7
kombu==2.5.7
lxml==3.0
mechanize==0.2.5
meld3==0.6.10
newrelic==1.9.0.21
numpy==1.7.0
oauth2==1.5.211
psycopg2==2.4.5
pylibmc==1.2.3
python-dateutil==1.5
python-openid==2.2.5
pytz==2012f
scikit-learn==0.13.1
scipy==0.11.0
simplejson==2.6.2
sqlitedict==1.0.9
supervisor==3.0b1
wsgiref==0.1.2
